Clean Entities and Beans
