+++ Tenzur/src/main/java/de/hhbk/beans/BeanTemplate.java
@@ -4,19 +4,11 @@ import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import java.io.Serializable;
 
-
 public abstract class BeanTemplate implements Serializable {
-    //-------------------------------------------------------------------------
-    //  Constructor(s)
-    //-------------------------------------------------------------------------
     public BeanTemplate() {
         super();
     }
 
-
-    //-------------------------------------------------------------------------
-    //  Else ...
-    //-------------------------------------------------------------------------
     protected void setMessage(String comonentId, FacesMessage.Severity type, String header, String msg) {
         FacesContext.getCurrentInstance().addMessage(comonentId, new FacesMessage(type, header, msg));
     }
@@ -28,6 +20,4 @@ public abstract class BeanTemplate implements Serializable {
     protected void setInfoMessage(String header, String msg) {
         setMessage(null, FacesMessage.SEVERITY_INFO, header, msg);
     }
-
-
 }
